Fixes types

+++ src/parserutils.c
@@ -2,6 +2,7 @@
 #include <limits.h>
 #include <stdlib.h>
 #include <string.h>
+#include <stddef.h>
 
 #include "parserutils.h"
 
@@ -37,11 +38,11 @@ bool safe_strtol(const char *str, long *out_val) {
     return true;
 }
 
-long get_operator_position(const char *formula) {
+ptrdiff_t  get_operator_position(const char *formula) {
     for (size_t i = 1, len = strlen(formula); i < len; i++) {
         char c = formula[i];
         if (c == '+' || c == '-' || c == '*' || c == '/') {
-            return i;
+            return (ptrdiff_t)i;
         }
     }
     return -1; // No operator found
@@ -49,7 +50,7 @@ long get_operator_position(const char *formula) {
 
 bool is_formula_valid(char *formula)
 {
-    long operator_position = get_operator_position(formula);
+    ptrdiff_t operator_position = get_operator_position(formula);
     if (operator_position == -1) {
         return false; // No operator found
     }
@@ -57,12 +58,12 @@ bool is_formula_valid(char *formula)
     for (size_t i = operator_position + 1, len = strlen(formula); i < len; i++) {
         char c = formula[i];
         if (c == '+' || c == '-' || c == '*' || c == '/') {
-            if (c != '-' || operator_position != (long)(i - 1)) {
+            if (c != '-' || operator_position != (ptrdiff_t)(i - 1)) {
                 return false; // More than one operator found or invalid operator sequence
             }
         }
     }
-    return operator_position != (long)(strlen(formula) - 1);
+    return operator_position != (ptrdiff_t)(strlen(formula) - 1);
 }
 
 bool try_get_operands(const char *formula, char **out_operand1, char **out_operand2, char *out_operator) {
@@ -70,7 +71,7 @@ bool try_get_operands(const char *formula, char **out_operand1, char **out_opera
         return false;
     }
 
-    int operator_position = get_operator_position(formula);
+    ptrdiff_t operator_position = get_operator_position(formula);
     
     // Extract operands and operator
     *out_operand1 = malloc(operator_position + 1);
